// ui/src/renderer/src/houston/generated/SwarmInfo.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SwarmSeverity } from "./SwarmSeverity";
import type { SwarmStatus } from "./SwarmStatus";

export type SwarmInfo = { id: number, name: string, root_dir: string, goal: string, status: SwarmStatus, created_at: number, budget_minutes: number, activated_at?: number | null, completed_at?: number | null, severity: SwarmSeverity, };